+++ CHANGELOG.md
@@ -5,8 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for compiling shaders in unsaved text editor
+
 ### Changed
 - Reusing shader output window to not spawn many windows
+- Optimized Windows SDK searching
+
+### Fixed 
+- Fix DXC was always used as shader compiler
 
 ## [1.0.0] - 2022-07-16
 ### Added